libft

+++ kayumi_cub3d/main.c
@@ -1,6 +1,6 @@
 #include "cub3d.h"
 
-size_t main(size_t argc, char *argv[])
+int main(int argc, char *argv[])
 {
 	int test;
 	test = set_map("original.txt");
